REFACTOR move speed units in subpackage
