+++ src/main/java/segelzwerg/sporttooolbox/iunits/speed/KilometerPerHour.java
@@ -1,8 +1,10 @@
-package segelzwerg.sporttooolbox.IUnits;
+package segelzwerg.sporttooolbox.iunits.speed;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
+import segelzwerg.sporttooolbox.IUnits.Speed;
+import segelzwerg.sporttooolbox.IUnits.Time;
 
 /**
  * Speed in kilometer per hour
