+++ src/main/java/segelzwerg/sporttooolbox/iunits/speed/MeterPerSecond.java
@@ -1,8 +1,10 @@
-package segelzwerg.sporttooolbox.IUnits;
+package segelzwerg.sporttooolbox.iunits.speed;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
+import segelzwerg.sporttooolbox.IUnits.Speed;
+import segelzwerg.sporttooolbox.IUnits.Time;
 
 /**
  * Speed in meter per second
