+++ src/test/java/segelzwerg/sporttooolbox/IUnits/DistanceTest.java
@@ -1,6 +1,7 @@
 package segelzwerg.sporttooolbox.IUnits;
 
 import org.junit.jupiter.api.Test;
+import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
