+++ src/test/java/segelzwerg/sporttooolbox/calculators/TimeCalculatorFactoryTest.java
@@ -2,8 +2,8 @@ package segelzwerg.sporttooolbox.calculators;
 
 import org.junit.jupiter.api.Test;
 import segelzwerg.sporttooolbox.IUnits.Distance;
-import segelzwerg.sporttooolbox.IUnits.KilometerPerHour;
 import segelzwerg.sporttooolbox.IUnits.Speed;
+import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
 import segelzwerg.sporttooolbox.web.speed.SpeedForm;
 
 import static org.hamcrest.MatcherAssert.assertThat;
