+++ src/test/java/segelzwerg/sporttooolbox/web/speed/SpeedServiceTest.java
@@ -2,9 +2,9 @@ package segelzwerg.sporttooolbox.web.speed;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import segelzwerg.sporttooolbox.IUnits.KilometerPerHour;
 import segelzwerg.sporttooolbox.IUnits.Speed;
 import segelzwerg.sporttooolbox.IUnits.Time;
+import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
